refactor(): more simplification and obsolete code removal * removed geometry generator variants * renamed buildCachedGeometryFromTreeDeep * updated the 'calling' code at various places
